+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   fixed time datatext error
+-   fixed tooltip anchor point
 
 ## [1.3.0] - 2024-10-15